Changed write method to 1 character only
Originally, this team had set up the number outputs to be two characters (base 256). However, this was (1) unable to be decoded [since null would print out, which was impossible to read] and (2) not the most efficient way (since there were two chars being used. I have switched it to 1 char being used, which goes all the way up to 65K+, greater than the max length established by this code.